Refactoring parte 1.
