+++ Emisor/Client/src/Controladores/ControladorVentanaMensaje.java
@@ -10,7 +10,7 @@ import Modelo.Mensaje.MensajeSimple;
 
 import Modelo.MensajesFactory;
 import Modelo.Sistema;
-import Modelo.Usuario;
+import Modelo.agenda.Usuario;
 
 import Vista.InterfazVistaMensaje;
