+++ Emisor/Client/src/Modelo/AgendaPersistencia.java
@@ -1,5 +1,7 @@
 package Modelo;
 
+import Modelo.agenda.Usuario;
+
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.IOException;
